feat: reduce the threshold for new product access.
