+++ src/components/ConversationList/index.tsx
@@ -1,7 +1,7 @@
 import { FC, useEffect } from 'react'
 import { useRecoilState, useRecoilValue } from 'recoil'
 import { db } from 'src/models/db'
-import { conversationTitles } from 'src/shared/constants'
+import { snakeCaseToTitleCase } from 'src/shared/utils'
 import { currConversationState } from 'src/stores/conversation'
 import { currProductState } from 'src/stores/global'
 import { Conversation } from 'src/types/conversation'
@@ -49,7 +49,7 @@ const ConversationList: FC<Props> = ({ conversations }) => {
     <section className="w-87.75">
       <section className="flex items-center justify-between p-6">
         <span className="mr-4 truncate text-xl font-semibold dark:text-dark-text">
-          {conversationTitles[currProduct]}
+          {snakeCaseToTitleCase(currProduct)}
         </span>
         <OutlinePlusIcon onClick={addConversation} />
       </section>
