+++ src/components/Sidebar/Items.tsx
@@ -18,7 +18,7 @@ import {
   OutlineTranslationIcon,
   SolidTranslationIcon
 } from 'src/components/Icons'
-import { conversationTitles } from 'src/shared/constants'
+import { snakeCaseToTitleCase } from 'src/shared/utils'
 import { Companies, Products } from 'src/types/global'
 
 export const iconClassName = 'h-6 w-6 text-black dark:text-white'
@@ -31,13 +31,13 @@ export default [
     products: [
       {
         product: Products.OpenAIChat,
-        tooltip: conversationTitles[Products.OpenAIChat],
+        tooltip: snakeCaseToTitleCase(Products.OpenAIChat),
         inactive: <ChatBubbleLeftRightIconOutline className={iconClassName} />,
         active: <ChatBubbleLeftRightIconSolid className={iconClassName} />
       },
       {
         product: Products.OpenAICompletion,
-        tooltip: conversationTitles[Products.OpenAICompletion],
+        tooltip: snakeCaseToTitleCase(Products.OpenAICompletion),
         inactive: (
           <ChatBubbleBottomCenterTextIconOutline className={iconClassName} />
         ),
@@ -47,19 +47,19 @@ export default [
       },
       {
         product: Products.OpenAIAudioTranscription,
-        tooltip: conversationTitles[Products.OpenAIAudioTranscription],
+        tooltip: snakeCaseToTitleCase(Products.OpenAIAudioTranscription),
         inactive: <MicrophoneIconOutline className={iconClassName} />,
         active: <MicrophoneIconSolid className={iconClassName} />
       },
       {
         product: Products.OpenAIAudioTranslation,
-        tooltip: conversationTitles[Products.OpenAIAudioTranslation],
+        tooltip: snakeCaseToTitleCase(Products.OpenAIAudioTranslation),
         inactive: <OutlineTranslationIcon className={iconClassName} />,
         active: <SolidTranslationIcon className={iconClassName} />
       },
       {
         product: Products.OpenAIImageGeneration,
-        tooltip: conversationTitles[Products.OpenAIImageGeneration],
+        tooltip: snakeCaseToTitleCase(Products.OpenAIImageGeneration),
         inactive: <PhotoIconOutline className={iconClassName} />,
         active: <PhotoIconSolid className={iconClassName} />
       }
@@ -71,13 +71,13 @@ export default [
     products: [
       {
         product: Products.AzureChat,
-        tooltip: conversationTitles[Products.AzureChat],
+        tooltip: snakeCaseToTitleCase(Products.AzureChat),
         inactive: <ChatBubbleLeftRightIconOutline className={iconClassName} />,
         active: <ChatBubbleLeftRightIconSolid className={iconClassName} />
       },
       {
         product: Products.AzureCompletion,
-        tooltip: conversationTitles[Products.AzureCompletion],
+        tooltip: snakeCaseToTitleCase(Products.AzureCompletion),
         inactive: (
           <ChatBubbleBottomCenterTextIconOutline className={iconClassName} />
         ),
@@ -87,7 +87,7 @@ export default [
       },
       {
         product: Products.AzureImageGeneration,
-        tooltip: conversationTitles[Products.AzureImageGeneration],
+        tooltip: snakeCaseToTitleCase(Products.AzureImageGeneration),
         inactive: <PhotoIconOutline className={iconClassName} />,
         active: <PhotoIconSolid className={iconClassName} />
       }
@@ -105,7 +105,7 @@ export default [
     products: [
       {
         product: Products.ClaudeChat,
-        tooltip: conversationTitles[Products.ClaudeChat],
+        tooltip: snakeCaseToTitleCase(Products.ClaudeChat),
         inactive: <ChatBubbleLeftRightIconOutline className={iconClassName} />,
         active: <ChatBubbleLeftRightIconSolid className={iconClassName} />
       }
